Remove ruby specification from Gemfile.
Unfortunately, Bundler requires the exact version specification which makes it
impossible to support Rubinius and MRI when this specification is use. This is
because Rubinius reports the Ruby version as 2.1.0 and does not change the teeny
version when MRI releases bug fixes because they are not Rubinius bugs. Ideally,
Bundler would allow the analogue of the RubyGems pessimistic operator and handle
eg '~> 2.1', or simply '2.1'.